fixed number of warnings; 2279 atomic changes...
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.3@4497 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/TEST/ModifierRep_test.C
@@ -124,7 +124,7 @@ START_SECTION(UInt getMaxModificationMasses())
 	ptr = new ModifierRep();
 	TEST_EQUAL(0,ptr->getMaxModificationMasses());
 	std::vector<std::vector<double> > mod_table = ptr->getModificationTable();
-	std::set<float> mod_masses_set;
+	std::set<double> mod_masses_set;
 	for (Size i = 0; i < mod_table.size();++i)
 	{
 		for (Size j = 0; j < mod_table.at(i).size();++j)
